EgoScript: Separate tokens from token decoders
