+++ egolib/egolib.vcxproj.filters
@@ -558,6 +558,9 @@
     <ClCompile Include="src\egolib\Script\PDLTokenKind.cpp">
       <Filter>Source Files\Script</Filter>
     </ClCompile>
+    <ClCompile Include="src\egolib\Script\DDLTokenDecoder.cpp">
+      <Filter>Source Files\Script</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\egolib\vfs.h">
@@ -1389,6 +1392,9 @@
       <Filter>Header Files\Script</Filter>
     </ClInclude>
     <ClInclude Include="src\egolib\Script\DDLTokenKind.hpp">
+      <Filter>Header Files\Script</Filter>
+    </ClInclude>
+    <ClInclude Include="src\egolib\Script\DDLTokenDecoder.hpp">
       <Filter>Header Files\Script</Filter>
     </ClInclude>
   </ItemGroup>
