+++ egolib/src/egolib/Script/DDLTokenDecoder.cpp
@@ -1,0 +1,1 @@
+#include "egolib/Script/DDLTokenDecoder.hpp"
